+++ rplugin/python3/denite/__init__.py
@@ -5,8 +5,9 @@
 # =============================================================================
 
 from importlib.util import find_spec
+from denite.rplugin import Rplugin
 
-if find_spec('vim'):
+if find_spec('yarp'):
     import vim
 elif find_spec('pynvim'):
     import pynvim
@@ -15,49 +16,36 @@ else:
     import neovim
     vim = neovim
 
+
 if hasattr(vim, 'plugin'):
     # Neovim only
-
-    from denite.ui.default import Default
 
     @vim.plugin
     class DeniteHandlers(object):
         def __init__(self, vim):
-            self._vim = vim
+            self._rplugin = Rplugin(vim)
 
         @vim.function('_denite_init', sync=True)
-        def init_python(self, args):
-            self._uis = {}
-            self._vim.vars['denite#_channel_id'] = self._vim.channel_id
-            return
+        def init_channel(self, args):
+            self._rplugin.init_channel(args)
 
         @vim.function('_denite_start', sync=True)
         def start(self, args):
-            try:
-                ui = self.get_ui(args[1]['buffer_name'])
-                return ui.start(args[0], args[1])
-            except Exception:
-                import traceback
-                import denite.util
-                for line in traceback.format_exc().splitlines():
-                    denite.util.error(self._vim, line)
-                denite.util.error(self._vim,
-                                  'Please execute :messages command.')
+            self._rplugin.start(args)
 
         @vim.function('_denite_do_action', sync=True)
         def take_action(self, args):
-            try:
-                ui = self.get_ui(args[0]['buffer_name'])
-                return ui._denite.do_action(args[0], args[1], args[2])
-            except Exception:
-                import traceback
-                import denite.util
-                for line in traceback.format_exc().splitlines():
-                    denite.util.error(self._vim, line)
-                denite.util.error(self._vim,
-                                  'Please execute :messages command.')
+            return self._rplugin.take_action(args)
 
-        def get_ui(self, buffer_name):
-            if buffer_name not in self._uis:
-                self._uis[buffer_name] = Default(self._vim)
-            return self._uis[buffer_name]
+if find_spec('yarp'):
+
+    global_denite = Rplugin(vim)
+
+    def _denite_init():
+        pass
+
+    def _denite_start(args):
+        global_denite.start(args)
+
+    def _denite_do_action(args):
+        return global_denite.do_action(args)
